Rename use-local-storage
